+++ packages/spotlight/CHANGELOG.md
@@ -2,7 +2,10 @@
 
 The following is a curated list of changes in the Enact spotlight module, newest changes on the top.
 
-## [unreleased]
+## [4.5.4] - 2023-06-07
+
+No significant changes.
+## [4.7.2] - 2023-07-14
 
 ### Fixed